+++ projects/ontimize-web-ngx/src/lib/config/o-providers.ts
@@ -17,6 +17,7 @@ import { Error403Component } from '../shared/components/error403/o-error-403.com
 import { O_MAT_ERROR_OPTIONS } from '../shared/material/o-mat-error/o-mat-error';
 import { Config } from '../types/config.type';
 import { Codes } from '../util/codes';
+import { Util } from '../util/util';
 
 function addPermissionsRouteGuard(injector: Injector) {
   const route = injector.get(Router);
@@ -34,8 +35,8 @@ export function appInitializerFactory(injector: Injector, config: Config, oTrans
       const storedLang = oTranslate.getStoredLanguage();
       const configLang = config['locale'];
       const browserLang = oTranslate.getBrowserLang();
-      let userLang = 'en';
-      let defaultLang = 'en';
+      let userLang = Util.isDefined(config['defaultLocale']) ? config['defaultLocale'] : 'en';
+      let defaultLang = Util.isDefined(config['defaultLocale']) ? config['defaultLocale'] : 'en';
       if (storedLang) {
         userLang = storedLang;
       } else if (configLang) {
@@ -45,14 +46,13 @@ export function appInitializerFactory(injector: Injector, config: Config, oTrans
         userLang = browserLang;
         defaultLang = browserLang;
       }
-
       oTranslate.setDefaultLang(defaultLang);
 
       const locales = new Set(config.applicationLocales || []);
       locales.add('en');
       locales.add(userLang);
 
-      // initialize available locales array if needed
+      // initialize available locales array if needed 
       if (!config.applicationLocales) {
         config.applicationLocales = [...locales];
       }